feat: moved map components to global components folder
