+++ src/components/map-component/index.ts
